fix: #WB-2608 lazyload katex css when editor display
